requirements.txt
tensorflow
numpy
matplotlib
streamlit
pillow
opencv-python
pandas